+++ src/boost_histogram/_internal/axestuple.py
