+++ tests/test_histogram_indexing.py
@@ -331,7 +331,7 @@ def test_pick_int_category():
 
 def test_axes_tuple():
     h = bh.Histogram(bh.axis.Regular(10, 0, 1))
-    assert isinstance(h.axes[:1], bh._internal.axistuple.AxesTuple)
+    assert isinstance(h.axes[:1], bh._internal.axestuple.AxesTuple)
     assert isinstance(h.axes[0], bh.axis.Regular)
 
     (before,) = h.axes.centers[:1]
@@ -344,7 +344,7 @@ def test_axes_tuple_Nd():
     h = bh.Histogram(
         bh.axis.Integer(0, 5), bh.axis.Integer(0, 4), bh.axis.Integer(0, 6)
     )
-    assert isinstance(h.axes[:2], bh._internal.axistuple.AxesTuple)
+    assert isinstance(h.axes[:2], bh._internal.axestuple.AxesTuple)
     assert isinstance(h.axes[1], bh.axis.Integer)
 
     b1, b2 = h.axes.centers[1:3]
